Support for IE

+++ aura-components/src/main/components/uiExamples/dragAndDropExample/dragAndDropExampleController.js
@@ -18,6 +18,7 @@
 		var context = component.find("list");
 		component.set("v._context1", context[0]);
 		component.set("v._context2", context[1]);
+		component.set("v._context3", context[2]);
 		component.set("v._accessibilityComponent", component.find("accessibilityComponent").getGlobalId());
 	},
 	
@@ -30,25 +31,28 @@
 		if (dropComplete) {
 			// Calculate index to be removed
 			var record = event.getParam("data");
+			var type = event.getParam("type");
 	    	var source = $A.dragAndDropService.getContext(event.getParam("dragComponent"));
 	    	var items = source.get("v.items");
 	    	
-	    	var removeIndex = -1;
-	    	for (var i = 0; i < items.length; i++) {
-				if (record === items[i]) {
-					removeIndex = i;
-					break;
-				}
+	    	if (type === "move") {
+		    	var removeIndex = -1;
+		    	for (var i = 0; i < items.length; i++) {
+					if (record === items[i]) {
+						removeIndex = i;
+						break;
+					}
+		    	}
+		    	
+				var removeParams = {
+					"index": removeIndex,
+					"count": 1,
+					"remove": true
+				};
+				
+				// Remove data transfer
+				$A.dragAndDropService.removeDataTransfer(event, removeParams);
 	    	}
-	    	
-			var removeParams = {
-				"index": removeIndex,
-				"count": 1,
-				"remove": true
-			};
-			
-			// Remove data transfer
-			$A.dragAndDropService.removeDataTransfer(event, removeParams);
 			
 			// Add data transfer
 			var target = $A.dragAndDropService.getContext(event.getParam("dropComponent"));
